requirements.txt
alembic==1.7.6
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.2
pandas==1.4.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
urllib3==1.26.8
zipp==3.7.0